Update UMIPs/umip-add rDAI as a collateral currency.md
Yes; I think this captures it. A "management fee" could be implemented either by setting one of the recipients in the hat to a contract sponsor's wallet, or (a little less transparently) by setting the excessTokenBeneficiary to the sponsor's wallet, and making the EMP contract one of the beneficiaries. I think it's more natural to give it a hat pointing to the charity; payInterest() through the rDAI interface is clearer than trimExcess() on the EMP

Co-authored-by: Sean Brown <sean.brown2796@gmail.com>