Change price feed from Uniswap to Bitfinex (#149)
After speaking with Sean, we realized price feeds aren't going to be able to natively support the ETH to USD normalization. We chose to submit an amendment to change the source to use the Bitfinex AMPLUSD price as it is high volume. Let me know if there are any questions regarding the proposed change.