improve(UMIP179): Omit 0-value refunds and 0-value empty-message slow fills
These two payments out of the spoke pool will do nothing on-chain and the dataworker should not produce merkle leafs containing them.

A slow fill could technically have an input amount of 0 but have a non-empty message, for example if the message pays the filler then it still might be deemed "fillable" by the filler.